remove notify_remaining_rate_limit
add optional root_url to initializer

we can remove the rate limiting since it depends on Logger and Honeybadger.  that logic can be added in
to the client that uses this.

adding an optional root_url will allow us to point to local or staged servers that "stand-in" for
livestorm to facilitate development without having to rely on the real service